Finishd the plot def. and  mainmenu still done.
have removed some funktions though. So if something is missing that is needed for the error funktion, we just get it from the old main. I can for some reason not create a new branch, so i just made a new file. 
Except for "check for error" this is ready to hand in.